refactor: split main.rs into focused app modules
main.rs had grown to 3191 lines holding provider setup, filter
construction, cache orchestration, and output rendering alongside main()
itself. Split it into src/app/{catalog,keys,selection,pipeline,caching,
report}.rs, moving each function's tests next to it. main.rs is now 281
lines of orchestration.

Removed the duplication the single-file layout had accumulated:

- apply_url_filters built a UrlFilter inline that was character-for-
  character the one build_url_filter returns, so the batch path and the
  --stream path could silently drift apart. Both now share one builder,
  as does the --extract-links filter.
- apply_url_transformations likewise duplicated build_url_transformer;
  it now reuses it and opts into --merge-endpoint, the single option
  only the batch path can honour.
- The four-line API-key parsing incantation appeared three times. It is
  now ApiKeys::resolve, and the per-provider flag/env-var names are
  derived from the provider id rather than hardcoded.
- The three "provider X requires an API key" errors are generated from
  the catalog, so a renamed flag can't leave a stale message behind.
- collect_domains and build_host_validator both reassembled the target
  list; they now share cli_domain_inputs.
- Tests carried three copies of a 70-field Args literal. They now share
  crate::test_support, which also holds the mocks and an EnvGuard that
  restores environment variables on unwind instead of leaking cleared
  state into later tests when an assertion panics.

Fixes a bug found while untangling process_domains_with_cache: under
--incremental, a domain that hit the cache was pushed onto the work list
twice -- once in the "still need fresh data" branch and again in the
fall-through that had no `continue` -- so every provider request for that
domain was issued twice for an identical result. Incremental runs now
skip the cache lookup entirely, since they always re-fetch.

Behaviour is otherwise unchanged: provider registration order, message
wording, and the stdin fallback (which keys off raw inputs, not
normalized ones, so an unparseable target can't send urx to a blocking
stdin read) are all preserved.

637 tests pass, up from 610; clippy -D warnings and fmt are clean.
